requirements.txt
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
dialogflow==0.4.0
google-api-core==1.2.1
google-auth==1.5.0
googleapis-common-protos==1.5.3
grpcio==1.12.1
idna==2.7
matrix-client==0.3.2
numpy==1.14.5
protobuf==3.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.1
pytz==2018.4
PyYAML==3.13
requests==2.19.1
rsa==3.4.2
six==1.11.0
urllib3==1.23
